Fix package structure for CLI client
