+++ client-cli/src/test/java/io/spine/examples/todolist/cli/ApplicationTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.todolist;
+package io.spine.examples.todolist.cli;
 
 import io.spine.cli.Application;
 import org.junit.jupiter.api.DisplayName;
