+++ client-cli/src/test/java/io/spine/examples/todolist/cli/view/ViewTest.java
@@ -18,10 +18,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.todolist.server.view;
+package io.spine.examples.todolist.cli.view;
 
-import io.spine.examples.todolist.AppConfig;
 import io.spine.examples.todolist.TodoListContext;
+import io.spine.examples.todolist.cli.AppConfig;
 import io.spine.examples.todolist.client.TodoClient;
 import io.spine.examples.todolist.server.Server;
 import io.spine.server.BoundedContext;
